% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backTrans-log.R
\name{backTrans_log}
\alias{backTrans_log}
\title{Back transform parameters estimated in the log domain}
\usage{
backTrans_log(.df, .col_list = c("value", "lower", "upper"))
}
\description{
Make sure any other calculations, such as CI (and pRSE) are
done before back-calculating these values.
}
\keyword{internal}
