% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backTrans_logit.R
\name{backTrans_logit}
\alias{backTrans_logit}
\title{Back transform parameters estimated in the logit domain}
\usage{
backTrans_logit(df)
}
\description{
Make sure any other calculations, such as CI (and pRSE) are
done before back-calculating these values.
}
