% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_param_table.R
\name{define_param_table}
\alias{define_param_table}
\title{Define parameter table}
\usage{
define_param_table(.estimates, .key)
}
\arguments{
\item{.estimates}{parameter estimates data.frame}

\item{.key}{parameter key}
}
\description{
Combines model output parameter estimates with information in parameter key. Performs
some formatting of this combined data.frame.

Expected input is a data.frame with parameter estimates, with the columns:
\code{parameter_names}, \code{estimate}, \code{stderr}, \code{random_effect_sd}, \code{random_effect_sdse},
\code{fixed}, \code{diag}, \code{shrinkage}.

Some \code{parameter_names} have punctuation such as \code{OMEGA(1,1)}. A new column is
added without punctuation, such as \code{OMEGA11}.

Following this, parameter details from the parameter key are joined to the parameter estimates.
A \code{dplyr::inner_join} is used so that only parameters in the model output are kept
in the table. This was done so that, if your base and final model used the same structural
THETAs and random parameters, the same parameter key could be used for both.

This join adds the following columns: \code{abb} (abbreviation), \code{desc} (parameter description),
\code{panel}, \code{trans} (transformation).

With this information provided, a check is performed to determine whether parameters
with special transformation rules were defined correctly. In addition, a series of
TRUE/FALSE columns that will be used by subsequent functions.
}
