% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineBootTable.R
\name{defineBootTable}
\alias{defineBootTable}
\title{Combine bootstrap estimates with non-bootstrap estimates and parameter key}
\usage{
defineBootTable(.boot_estimates, .nonboot_estimates, .key)
}
\arguments{
\item{.boot_estimates}{parameter boot estimates- either path to file or data.frame}

\item{.nonboot_estimates}{non-bootstrap final model - either path to file or bbr model_summary}

\item{.key}{parameter key- either path to file or data.frame}
}
\description{
Combines boot strap estimates and non boot strap estimates with information in parameter key. Performs
some formatting of this combined data.frame.
There are two main steps of this function:
\enumerate{
\item Run \code{bbr::param_estimates_compare} to extract summary quantiles, the 5th, 50th, and 95th, of the
bootstrap estimates for each model parameter.
}

Some \code{parameter_names} have punctuation such as \code{OMEGA(1,1)}. A new column is
added without punctuation, such as \code{OMEGA11}.

Following this, parameter details from the parameter key are joined to the boot strap parameter estimates.
A \code{dplyr::inner_join} is used so that only parameters in the model output are kept
in the table. This was done so that, if your base and final model used the same structural
THETAs and random parameters, the same parameter key could be used for both.

This join adds the following columns: \code{abb} (abbreviation), \code{desc} (parameter description),
\code{panel}, \code{trans} (transformation).
\enumerate{
\item Reformat non-bootstrap estimates and left join onto combined bootstrap estimates and parameter key data.frame.
Expected input is a data.frame with parameter estimates, with the columns:
\code{parameter_names}, \code{estimate}.
}

With this information provided, a check is performed to determine whether parameters
with special transformation rules were defined correctly. In addition, a series of
TRUE/FALSE columns that will be used by subsequent functions.
}
\examples{

boot_paramEst <- utils::read.csv(system.file("model/nonmem/boot/data/boot-106.csv",
                                 package = "mrgparamtab"))
nonboot_paramEst <- utils::read.csv(system.file("model/nonmem/nonboot_param_est.csv",
                                    package = "mrgparamtab"))
paramKey <-  system.file("model/nonmem/pk-parameter-key-new.yaml", package = "mrgparamtab")

defineBootTable(.boot_estimates = boot_paramEst,
               .nonboot_estimates = nonboot_paramEst,
               .key = paramKey)

}
\seealso{
\link[mrgparamtab]{param_key}: Parameter key requirements
}
