% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineRows.R
\name{defineRows}
\alias{defineRows}
\title{Define a series of true/false columns}
\usage{
defineRows(.df)
}
\arguments{
\item{.df}{data frame containing parameter estimates & names}
}
\description{
Creates variables with true/false values depending on the type of parameter
on each row. For instance, an "OMEGA" parameter would be OM ~ TRUE and TH ~ FALSE & S ~ FALSE.

Similarly, true/false variables are created to indicate the transform selected.
}
\keyword{internal}
