% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatValues.R
\name{display_value}
\alias{display_value}
\title{Format values for display purposes}
\usage{
display_value(x, .digit = NULL, .maxex = NULL)
}
\arguments{
\item{x}{Numeric; value to manipulate}

\item{.digit}{Number of significant digits. Default is three digits}

\item{.maxex}{Maximum number of significant digits before moving to scientific
notation. Default is \code{NULL}}
}
\description{
Format significant digits and substitute negative signs (\code{-}) for LaTeX
styling (\verb{$-$})
}
\keyword{internal}
