% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_boot_table.R
\name{format_boot_cols}
\alias{format_boot_cols}
\title{Format and group bootstrap \code{perc_[x]} columns}
\usage{
format_boot_cols(.df)
}
\arguments{
\item{.df}{bootstrap parameter dataset and non-bootstrap parameter dataset combined.}
}
\description{
Format and group bootstrap \code{perc_[x]} columns. Columns that correspond to a
confidence interval will instead be grouped into a new confidence interval
column displaying the range. E.g., \code{perc_2.5} + \code{perc_97.5} --> \verb{95\% CI}
}
\keyword{internal}
