% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-param-table.R
\name{format_param_table}
\alias{format_param_table}
\title{Format parameter table}
\usage{
format_param_table(
  .df,
  .select_cols = c("type", "abb", "greek", "desc", "value", "ci", "shrinkage")
)
}
\arguments{
\item{.df}{parameter estimates with modifications ready for formatting}

\item{.select_cols}{columns to select for output}
}
\description{
Format parameter estimate values and output selected columns to be shown in
the parameter table.
}
