% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatGreekNames.R
\name{formatGreekNames}
\alias{formatGreekNames}
\title{Format THETA/OMEGA/SIGMA values}
\usage{
formatGreekNames(df)
}
\description{
Format THETA/OMEGA/SIGMA values to display as greek letters with subscript numbers.
}
