% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get90CI.R
\name{get90CI}
\alias{get90CI}
\title{Calculate 90\% confidence intervals}
\usage{
get90CI(.df, .value = "value", .se = "se")
}
\arguments{
\item{.value}{column name to be used to calculate confidence interval}

\item{.se}{column name with standard errors}

\item{df}{data.frame with parameter estimates}
}
\description{
Calculates the upper and lower bound of a 95\% confidence interval, based on the
value and standard error.
}
\examples{

Theoph_ex <- Theoph \%>\% dplyr::mutate(stderr = conc/10)
get95CI(.df = Theoph_ex, .value = "conc", .se = "stderr")

}
