% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCV-propS.R
\name{getCV_propS}
\alias{getCV_propS}
\title{Get CV proportion}
\usage{
getCV_propS(v)
}
\description{
CV equations from \href{https://ascpt.onlinelibrary.wiley.com/doi/full/10.1002/psp4.12404}{NONMEM Tutorial Part I: Description of Commands and Options, With Simple Examples of Population Analysis}
}
\keyword{internal}
