% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpCV.R
\name{getpCV}
\alias{getpCV}
\title{Calculate CV\%}
\usage{
getpCV(.df, .column = "value")
}
\arguments{
\item{.df}{data.frame with combined parameter estimates and parameter key}

\item{.column}{column name to perform calculation on}
}
\description{
Calculate CV\%
}
