% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValueSE.R
\name{getValueSE}
\alias{getValueSE}
\title{Determine a value (or estimate) for each model parameter and it’s associated metric}
\usage{
getValueSE(.df)
}
\arguments{
\item{.df}{data.frame with parameter estimates}
}
\description{
value should have estimate (something):

theta = estimate only                           # use estimate column

omega diagonals = variance (\%CV)                # estimate (CV from estimate, stderr)

omega off-diagonals = covariance (corr coeff)   # estimate (random_effect_sd)

sigma diagonal proportional = variance (\%CV)    # estimate (CV from estimate, stderr)

sigma diagonal additive = variance (SD)         # estimate (random_effect_sd)
}
\keyword{internal}
