% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupdf-check.R
\name{grouped_warn_ungroup}
\alias{grouped_warn_ungroup}
\title{Confirm data frames are not grouped}
\usage{
grouped_warn_ungroup(.df, .func)
}
\description{
Returns a warning if an input dataframe is grouped when given to a function not
anticipating a grouped column.
}
\keyword{internal}
