% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadParamEstimates.R
\name{loadParamEstimates}
\alias{loadParamEstimates}
\title{Load parameter estimates}
\usage{
loadParamEstimates(.estimates)
}
\arguments{
\item{.estimates}{path to model directory, bbr NONMEM model, or data.frame of parameter estimates}
}
\description{
Loads parameter estimates by providing model path, bbr model summary, or data frame of parameter estimates
}
\keyword{internal}
