% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadParamKey.R
\name{loadParamKey}
\alias{loadParamKey}
\title{Load parameter key}
\usage{
loadParamKey(.key)
}
\arguments{
\item{.key}{path to parameter key or data.frame of parameter key}
}
\description{
Loads parameter key by providing parameter key path  or data frame of parameter key.
Checks that parameter key contains the following columns: name, abb, desc, panel, trans.
}
\keyword{internal}
