% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_int.R
\name{lower95CI}
\alias{lower95CI}
\alias{upper95CI}
\alias{lower90CI}
\alias{upper90CI}
\title{Confidence interval bounds}
\usage{
lower95CI(est, se)

upper95CI(est, se)

lower90CI(est, se)

upper90CI(est, se)
}
\arguments{
\item{est}{parameter estimates}

\item{se}{standard error}
}
\description{
Confidence interval bounds
}
\keyword{internal}
