% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_int.R
\name{lowerCI}
\alias{lowerCI}
\alias{upperCI}
\title{Confidence interval bounds}
\usage{
lowerCI(.est, .se, .ci)

upperCI(.est, .se, .ci)
}
\arguments{
\item{.est}{parameter estimates}

\item{.se}{standard error}

\item{.ci}{confidence interval type}
}
\description{
Confidence interval bounds
}
\keyword{internal}
