% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-boot-pmtable.R
\name{make_boot_pmtable}
\alias{make_boot_pmtable}
\title{Make pmtable boot parameter table}
\usage{
make_boot_pmtable(.boot_df, .nonboot_df, .pmtype = "full", .width = 1)
}
\arguments{
\item{.boot_df}{bootstrap parameter data set output from pmparams::format_boot_table.}

\item{.nonboot_df}{parameter data set output from pmparams::format_param_table. Default is NULL.}

\item{.pmtype}{parameter table type. Options include: full (all rows in .df retained in pmtable), fixed (all rows with type = "Struct" or "effect"), structural (all rows with type = "Struct"), covariate (all rows with type = "effect"), random (all rows with greek = "Omega" or type = "Resid"). Defaults to "full".}

\item{.width}{notes width. Defaults to 1.}
}
\description{
Generates specific parameter tables by filtering and using pmtables

This function:
\enumerate{
\item Filters to columns needed for specific parameter tables
\item Panels by "type"
\item Makes "abb", "greek", "desc" blank for "full", "fixed", "structural", "covariate" parameter tables.
Makes "abb" and "greek" blank for "random"
\item Attaches notes
\item Rename "value" to "Estimate" and "shrinkage" to "Shrinkage (\\\%)", if applicable
}

If these pmtable settings do not work for your parameter table, you can overwrite them afterwards using desired pmtables commands.
}
\examples{

#Using output from `format_param_table` (defineOut),

#To make random effects table:

make_boot_pmtable(.df = data, .pmtype = "random")
}
