% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greek-num.R
\name{mathMode}
\alias{mathMode}
\alias{gtGreek}
\alias{greekNum}
\alias{expGreek}
\alias{logitGreek}
\title{Greek number helper functions}
\usage{
mathMode(.x)

gtGreek(.x)

greekNum(.x, .y)

expGreek(.x, .y)

logitGreek(.x, .y)
}
\description{
Greek number helper functions
}
\keyword{internal}
