% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-key.R
\name{param_key}
\alias{param_key}
\title{Parameter key requirements}
\description{
The parameter key used in the pmparams functions can either be a yaml or data frame.
It must contain the following:

\strong{name} or \strong{yaml variable}= THETA, OMEGA, SIGMA. See examples below.

\strong{abb} = abbreviation you want to appear in the parameter table (use latex coding)

\strong{desc} = description you want to appear in the parameter table

\strong{panel} = indicate which panel each parameter should appear under.

Current options include:
\itemize{
\item \code{panel=="struct"} ~ "Structural model parameters"
\item \code{panel=="cov"} ~ "Covariate effect parameters"
\item \code{panel=="IIV"} ~ "Interindividual covariance parameters" (off-diagonals << function takes care of this)
\item \code{panel=="IIV"} ~ "Interindividual variance parameters" (diagonals << function takes care of this)
\item \code{panel=="IOV"}  ~ "Interoccasion variance parameters"
\item \code{panel=="RV"} ~ "Residual variance"
}

\strong{trans} = define how the parameter will be transformed. Current options include:
\itemize{
\item \code{"none"} -   untransformed parameters, e.g. THETAs or off-diagonals
\item \code{"logTrans"} -  THETAs estimated in the log domain
\item \code{"logitTrans"} - THETAs estimated using a logit transform
\item \code{"lognormalOm"}- for log-normal omegas e.g. \code{CL = THETA(1) * exp(ETA(1))} - returns est.+CV\%
\item \code{"OmSD"}       - for omegas where you want to return SD only - returns estimate & SD; use this for additive omegas e.g. \code{CL = THETA(1) + ETA(1)}
\item \code{"logitOmSD"}  - for omegas using logit transform - returns estimate & SD (calculated with \code{logitnorm} package); this option requires you provide the associated \code{THETA} separated with a "~"; e.g. \code{"logitOmSD ~ THETA3"}
\item \code{"addErr"}     - for additive error terms (coded using \code{THETA} or \code{SIGMA} in \verb{$ERROR}) - returns est.+ SD
\item \code{"propErr"}    - for proportional error terms (coded using \code{THETA} or \code{SIGMA} in \verb{$ERROR}) - returns est.+CV\%
\item \code{"addErrLogDV"} - for an additive error term on log-transformed data (coded using \code{THETA} or \code{SIGMA} in \verb{$ERROR}) - returns est.+CV\%
}
}
\section{YAML Example}{
Example of YAML syntax for parameter key file

\if{html}{\out{<div class="sourceCode">}}\preformatted{THETA1
  abb: "KA (1/h)"
  desc: "First order absorption rate constant"
  panel: struct
  trans: logTrans

THETA2
  abb: "V2/F (L)"
  desc: "Apparent central volume"
  panel: struct
  trans: logTrans
}\if{html}{\out{</div>}}
}

\examples{

# Example of parameter key data frame:
# It must contain parameter names, abb, desc, panel, and trans columns.

paramKey = dplyr::tribble(
  ~name, ~abb, ~desc, ~panel, ~trans,
  "THETA1",  "KA (1/h)", "First order absorption rate constant",   "struct", "logTrans",
  "THETA2", "V2/F (L)",  "Apparent central volume",                "struct", "logTrans"
)
paramKey

}
