% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-notes.R
\name{param_notes}
\alias{param_notes}
\title{Generate all footnote equation options}
\usage{
param_notes(.ci = 95, .zscore = NULL)
}
\arguments{
\item{.ci}{Confidence interval. A value from 1 to 99 denoting the percent
confidence interval. Default is \code{95}.}

\item{.zscore}{Deprecated. Please use the \code{.ci} argument.}
}
\description{
Generate list of generic footnote equations to append to parameter tables.
}
\examples{
model_dir <- system.file("model/nonmem", package = "pmparams")
paramKey <-  file.path(model_dir, "pk-parameter-key-new.yaml")

# Read in parameter estimates (or provide `bbr` model)
param_est <- utils::read.csv(file.path(model_dir, "param_est_102.csv"))

# Make and format parameter table
param_df <- param_est \%>\%
  define_param_table(paramKey, .ci = 95) \%>\%
  format_param_table()

# Make random effects table and add relevant footnotes:
footnotes <- param_notes(.ci = 95)

table <- param_df \%>\%
  make_pmtable(.pmtype = "random") \%>\%
  pmtables::st_notes(footnotes$ci, footnotes$cv) \%>\%
  pmtables::st_notes_str() \%>\%
  pmtables::st_notes(footnotes$ciEq) \%>\%
  pmtables::st_notes(footnotes$cvOmegaEq, footnotes$cvSigmaEq)

}
\seealso{
\code{\link[=boot_notes]{boot_notes()}}
}
