% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parens.R
\name{parens}
\alias{parens}
\alias{parensSQ}
\alias{parensSQ_CV}
\alias{parensSQ_corr}
\alias{parensSQ_se}
\title{Add parentheses around values}
\usage{
parens(.x)

parensSQ(.x)

parensSQ_CV(.x)

parensSQ_corr(.x)

parensSQ_se(.x)
}
\arguments{
\item{.x}{value to add parentheses around}
}
\description{
Add parentheses around values
}
\keyword{internal}
