% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_quantiles.R
\name{qt}
\alias{qt}
\alias{med}
\alias{lo}
\alias{hi}
\title{Calculate quartiles when using bootstrap estimates}
\usage{
qt(.x, .prob)

med(.x)

lo(.x)

hi(.x)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted}

\item{prob}{numeric vector of probabilities with values in \link{0,1}}
}
\description{
Calculate quartiles when using bootstrap estimates
}
\keyword{internal}
