% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_ci_msg.R
\name{zscore_ci_msg}
\alias{zscore_ci_msg}
\title{Generate message that confidence interval and z-score do not match}
\usage{
zscore_ci_msg(.ci, .zscore)
}
\description{
Output message if confidence interval and z-score do not match for \code{get_equations}
}
\keyword{internal}
