% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_boot_table.R
\name{format_boot_table}
\alias{format_boot_table}
\title{Format bootstrap parameter table}
\usage{
format_boot_table(
  .boot_df,
  .select_cols = c("abb", "desc", "boot_value", "boot_ci"),
  .digit = getOption("pmparams.dig"),
  .maxex = getOption("pmparams.maxex")
)
}
\arguments{
\item{.boot_df}{parameter estimates output from \code{define_boot_table} with modifications ready for formatting}

\item{.select_cols}{columns to select for output. Default selects "abb", "desc", "boot_value", "boot_ci". To return all columns, specify "all" for .select_cols.}

\item{.digit}{set significant digits for output (optional). Default is three digits}

\item{.maxex}{set maxex for computation (optional). Default is NULL}
}
\description{
Format bootstrap parameter estimate values and output selected columns to be shown in
the bootstrap parameter table.

Left join this output bootstrap data.frame with the formatted output of non-bootstrap parameter estimates
}
\examples{

#Using output from `define_boot_table` (defineBootOut),
boot_paramEst <- utils::read.csv(system.file("model/nonmem/boot/data/boot-106.csv",
                                 package = "pmparams"))
nonboot_paramEst <- utils::read.csv(system.file("model/nonmem/nonboot_param_est.csv",
                                    package = "pmparams"))
paramKey <-  system.file("model/nonmem/pk-parameter-key-new.yaml", package = "pmparams")

defineBootOut <- define_boot_table(.boot_estimates = boot_paramEst,
               .nonboot_estimates = nonboot_paramEst,
               .key = paramKey)

format_boot_table(.boot_df = defineBootOut)

#To include all columns:

format_boot_table(.boot_df = defineBootOut, .select_cols="all")
}
