% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_boot_table.R
\name{format_boot_table}
\alias{format_boot_table}
\title{Format bootstrap parameter table}
\usage{
format_boot_table(
  .boot_df,
  .select_cols = c("abb", "desc", "boot_value", "boot_ci_95"),
  .digit = getOption("pmparams.dig"),
  .maxex = getOption("pmparams.maxex")
)
}
\arguments{
\item{.boot_df}{parameter estimates output from \code{\link[pmparams]{define_boot_table}} with modifications ready for formatting}

\item{.select_cols}{columns to select for output. Default selects "abb", "desc", "boot_value", "boot_ci". To return all columns, specify "all" for .select_cols.}

\item{.digit}{set significant digits for output (optional). Default is three digits}

\item{.maxex}{set maxex for computation (optional). Default is NULL}
}
\description{
Format bootstrap parameter estimate values and output selected columns to be shown in
the bootstrap parameter table.

Left join this output bootstrap data.frame with the formatted output of non-bootstrap parameter estimates
}
\examples{
model_dir <- system.file("model/nonmem", package = "pmparams")

# Using output from `define_boot_table` (defineBootOut),
boot_paramEst <- readr::read_csv(file.path(model_dir, "boot/data/boot-106.csv"))
nonboot_paramEst <- readr::read_csv(file.path(model_dir, "param_est_106.csv"))
paramKey <-  file.path(model_dir, "pk-parameter-key-new.yaml")

defineBootOut <- define_boot_table(.boot_estimates = boot_paramEst,
               .nonboot_estimates = nonboot_paramEst,
               .key = paramKey)

format_boot_table(.boot_df = defineBootOut)

# To include all columns:

format_boot_table(.boot_df = defineBootOut, .select_cols="all")
}
