% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_param_table.R
\name{format_param_table}
\alias{format_param_table}
\title{Format parameter table}
\usage{
format_param_table(
  .df,
  .cleanup_cols = TRUE,
  .prse = FALSE,
  .digit = getOption("pmparams.dig"),
  .maxex = getOption("pmparams.maxex"),
  .select_cols = NULL
)
}
\arguments{
\item{.df}{parameter estimates output from \code{\link[=define_param_table]{define_param_table()}} with
modifications ready for formatting.}

\item{.cleanup_cols}{logical (T/F). Defaults to \code{TRUE}, which selects the
following columns:
\itemize{
\item \code{"type"}, \code{"abb"}, \code{"greek"}, \code{"desc"}, \code{"value"}, \code{"ci"}, \code{"shrinkage"}.
\item Set to \code{FALSE} to return all columns.
}}

\item{.prse}{logical (T/F). If \code{TRUE}, output \code{pRSE}. Default is \code{FALSE}.}

\item{.digit}{set significant digits for output (optional). Default is three
digits.}

\item{.maxex}{set maxex for computation (optional). Default is \code{NULL}.}

\item{.select_cols}{Deprecated. Please use \code{.cleanup_cols} instead. Columns
to select for output. To return all columns, specify \code{"all"}.}
}
\description{
Format parameter estimate values and output selected columns to be shown in
the parameter table.
}
\details{
There are four main steps of this function:
\enumerate{
\item When necessary, back transform parameters and their CIs, round parameters
using \code{pmtables::sig}, and combine columns.
\item Format the THETA/OMEGA/SIGMA values to display as Greek letters in LaTeX,
with subscript numbers, and indicate the transformation applied to that
parameter, if applicable.
\item Determine which panel of the final table the parameter should be displayed
in. This is informed by the panel argument you defined in your parameter key.
Note that there are a finite number of options included by default (see
below), but you can include additional panels as needed.

Panel types include:
\itemize{
\item Residual variance
\item Interindividual covariance parameters
\item Interindividual variance parameters
\item Interoccasion variance parameters
\item Covariate effect parameters
\item Structural model parameters
}
\item Select columns for final tables.
}
}
\examples{

model_dir <- system.file("model/nonmem", package = "pmparams")
paramKey <-  file.path(model_dir, "pk-parameter-key-new.yaml")

# Using output from `define_param_table` (param_df),
mod <- bbr::read_model(file.path(model_dir, "106"))
define_param_table(
 .estimates = mod,
 .key = paramKey,
 .ci = 95,
)

format_param_table(param_df)

# To include all columns:
format_param_table(param_df, .cleanup_cols = FALSE)
}
