% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatBootTable.R
\name{formatBootTable}
\alias{formatBootTable}
\title{Format bootstrap parameter table}
\usage{
formatBootTable(
  .boot_df,
  .select_cols = c("abb", "desc", "boot_value", "boot_ci"),
  .digit = getOption("mrgparamtab.dig"),
  .maxex = getOption("mrgparamtab.maxex")
)
}
\arguments{
\item{.boot_df}{parameter estimates output from \code{defineBootTable} with modifications ready for formatting}

\item{.select_cols}{columns to select for output. Default selects "abb", "desc", "boot_value", "boot_ci". To return all columns, specify "all" for .select_cols.}

\item{.digit}{set significant digits for output (optional)}

\item{.maxex}{set maxex for computation (optional)}
}
\description{
Format bootstrap parameter estimate values and output selected columns to be shown in
the bootstrap parameter table.

Left join with data.frame with the formatted output of non-bootstrap parameter estimates
}
