% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatParamTable.R
\name{formatParamTable}
\alias{formatParamTable}
\title{Format parameter table}
\usage{
formatParamTable(
  .df,
  .select_cols = c("type", "abb", "greek", "desc", "value", "ci", "shrinkage")
)
}
\arguments{
\item{.df}{parameter estimates with modifications ready for formatting}

\item{.select_cols}{columns to select for output}
}
\description{
Format parameter estimate values and output selected columns to be shown in
the parameter table.
}
