% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatValues.R
\name{formatValues}
\alias{formatValues}
\title{Define how values are to be displayed}
\usage{
formatValues(df)
}
\description{
Define what is in estimate column and what is in square brackets.

Displays 95\% CI to show lower, upper or FIXED and rounds values for display in report
}
\keyword{internal}
