% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatValuesBoot.R
\name{formatValuesBoot}
\alias{formatValuesBoot}
\title{Define how values are to be displayed}
\usage{
formatValuesBoot(.boot_df, .digit, .maxex)
}
\description{
Format values for bootstrap run
}
\keyword{internal}
