% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_equations.R
\name{get_equations}
\alias{get_equations}
\title{Generate all footnote equation options}
\usage{
get_equations(.ci = 95, .zscore = NULL)
}
\arguments{
\item{.ci}{specify 90 or 95 percent confidence interval (default 95\%)}

\item{.zscore}{z-score for the specified confidence interval. Only needed for confidence intervals that are NOT 90 or 95 percent}
}
\description{
Generate data frame of generic footnote equations to append to parameter tables.
}
\examples{
#Using output from `format_param_table` (defineOut),
library(dplyr)
paramEst <- utils::read.csv(system.file("model/nonmem/param_est.csv", package = "pmparams"))
paramKey <-  system.file("model/nonmem/pk-parameter-key-new.yaml", package = "pmparams")
defineOut <- define_param_table(.estimates = paramEst, .key = paramKey, .ci = 95, .zscore = NULL)
data <- format_param_table(.df = defineOut)

#To make random effects table and add relevant footnotes:
footnotes = get_equations()
table <- make_pmtable(.df = data, .pmtype = "random") \%>\%
pmtables::st_notes(footnotes$ci, footnotes$cv) \%>\%
pmtables::st_notes_str() \%>\%
pmtables::st_notes(footnotes$cvOmegaEq, footnotes$cvSigmaEq)

}
