% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBootPercentiles.R
\name{getBootPercentiles}
\alias{getBootPercentiles}
\title{Calculate percentiles}
\usage{
getBootPercentiles(.boot, .ci = 95, .na.rm = TRUE)
}
\arguments{
\item{.boot}{data.frame with parameter estimates}

\item{.ci}{the confidence interval. A value from 1 to 100 denoting the percent
confidence interval, or \code{"iqr"} (interquartile range). Default is \code{95}.}

\item{.na.rm}{logical (T/F). If \code{TRUE}, any \code{NA} and \code{NaN}'s are removed
before the quantiles are computed. Passed to \code{\link[stats:quantile]{stats::quantile()}}.}
}
\description{
Calculates the upper and lower bound of a 90\% or 95\% confidence interval,
interquartile range, or any number of percentiles based on the value and
standard error.
}
\keyword{internal}
