% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBootPercentiles.R
\name{getBootPercentiles}
\alias{getBootPercentiles}
\title{Calculate percentiles}
\usage{
getBootPercentiles(.boot, .ci, .percentiles, .na.rm)
}
\arguments{
\item{.ci}{specify IQR, 90 or 95 confidence intervals. Default is 95.}

\item{.percentiles}{list of all percentiles. Default is NULL. For .ci = 95, .percentiles = c(0.025, 0.5, 0.975). For .ci =90, .percentiles will be c(0.05, 0.5, 0.95)}

\item{.na.rm}{Default is false}

\item{.df}{data.frame with parameter estimates}
}
\description{
Calculates the upper and lower bound of a 90\% or 95\% confidence interval, interquartile range, or any number of percentiles based on the
value and standard error.
}
\keyword{internal}
