% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCI.R
\name{getCI}
\alias{getCI}
\title{Calculate 90\% or 95\% confidence intervals}
\usage{
getCI(.df, .value = "value", .se = "se", .ci = 95)
}
\arguments{
\item{.df}{data.frame with parameter estimates}

\item{.value}{column name to be used to calculate confidence interval}

\item{.se}{column name with standard errors}

\item{.ci}{specify 90 or 95 percent confidence interval (default 95\%)}

\item{.zscore}{z-score for the specified confidence interval. Only needed for confidence intervals that are NOT 90 or 95 percent}
}
\description{
Calculates the upper and lower bound of a 90\% or 95\% confidence interval, based on the
value and standard error.
}
\keyword{internal}
