% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCI.R
\name{getCI}
\alias{getCI}
\title{Calculate 90\% or 95\% confidence intervals}
\usage{
getCI(.df, .value = "value", .se = "se", .ci = 95)
}
\arguments{
\item{.df}{A data.frame with parameter estimates}

\item{.value}{Column name to be used to calculate confidence interval}

\item{.se}{Column name with standard errors}

\item{.ci}{Confidence interval. A value from 1 to 99 denoting the percent
confidence interval. Default is \code{95}.}
}
\description{
Calculates the upper and lower bound of a confidence interval, based on the
value and standard error.
}
\keyword{internal}
