% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpRSE.R
\name{getpRSE}
\alias{getpRSE}
\title{Calculate \% RSE}
\usage{
getpRSE(
  .df,
  .value = "value",
  .se = "se",
  ...,
  .digit = getOption("mrgparamtab.dig"),
  .maxex = getOption("mrgparamtab.maxex")
)
}
\arguments{
\item{.df}{data.frame with parameter estimates}

\item{.value}{column name of value of interest}

\item{.se}{column name containing standard error}

\item{.digit}{set significant digits for output (optional)}

\item{.maxex}{set maxex for computation (optional)}
}
\description{
Calculate \% RSE

Note, this is appropriate when parameters are estimated untransformed or in the log.
It may not be appropriate if any other transformations (such as logit) were used
}
\keyword{internal}
