% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpRSE.R
\name{getpRSE}
\alias{getpRSE}
\title{Calculate \% RSE}
\usage{
getpRSE(
  .df,
  .digit = getOption("mrgparamtab.dig"),
  .maxex = getOption("mrgparamtab.maxex")
)
}
\description{
Note, this is appropriate when parameters are estimated untransformed or in the log.
It may not be appropriate if any other transformations (such as logit) were used
}
