% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSD_logit0.R, R/getSD_logitO.R
\name{getSD_logitO}
\alias{getSD_logitO}
\title{SD and \%CV for logit-normal distributions}
\usage{
getSD_logitO(.mean, .var, .a = 0, .b = 1)

getSD_logitO(.mean, .var, .a = 0, .b = 1)
}
\arguments{
\item{.mean}{mean of the logit term}

\item{.var}{variance of the logit term}

\item{.a}{additive term}

\item{.b}{proportional term}
}
\description{
CV\% not appropriate for logit transforms so report SD
\%CV for random variable Y.

Y = a + b * (1 / (1 + exp(-X)))

Where X ~ N(mu, sigma) is the normally-distributed logit term
e.g. for PARAM = 1 / (1 + EXP(-(THETA1 + ETA1)))

CV equations from \href{https://ascpt.onlinelibrary.wiley.com/doi/full/10.1002/psp4.12404}{NONMEM Tutorial Part I: Description of Commands and Options, With Simple Examples of Population Analysis}

Decided CV\% not appropriate for logit transforms so report SD
\%CV for random variable Y.

Y = a + b * (1 / (1 + exp(-X)))

Where X ~ N(mu, sigma) is the normally-distributed logit term
e.g. for PARAM = 1 / (1 + EXP(-(THETA1 + ETA1)))
}
