% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValueSE.R
\name{getValueSE}
\alias{getValueSE}
\title{Determine a value (or estimate) for each model parameter and it’s associated metric}
\usage{
getValueSE(
  .df,
  .digit = getOption("mrgparamtab.dig"),
  .maxex = getOption("mrgparamtab.maxex")
)
}
\arguments{
\item{.df}{data.frame with parameter estimates}

\item{.digit}{set the number of significant digits}

\item{.maxex}{set the number of maxex}
}
\description{
value should have estimate \link{something}:

theta = estimate only                           # use estimate column

omega diagonals = variance \link{\%CV}                # estimate \link{CV from estimate, stderr}

omega off-diagonals = covariance \link{corr coeff}   # estimate \link{random_effect_sd}

sigma diagonal proportional = variance \link{\%CV}    # estimate \link{CV from estimate, stderr}

sigma diagonal additive = variance \link{SD}         # estimate \link{random_effect_sd}
}
