% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadParamEstimates.R
\name{loadParamEstimates}
\alias{loadParamEstimates}
\title{Load parameter estimates}
\usage{
loadParamEstimates(.estimates)
}
\arguments{
\item{.estimates}{One of the following:
\itemize{
\item Path to a \code{bbr} model
\item A \code{bbr} model. (output of \code{bbr::read_model()})
\item A \code{bbr} model summary object (output of \code{bbr::model_summary()})
\item A data frame of the parameter estimates (e.g., output of \code{bbr::param_estimates()})
}}
}
\description{
Loads parameter estimates by providing model path, bbr model summary, or data frame of parameter estimates
}
\keyword{internal}
