% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_boot_pmtable.R
\name{make_boot_pmtable}
\alias{make_boot_pmtable}
\title{Make pmtable bootstrap parameter table}
\usage{
make_boot_pmtable(
  .df,
  .pmtype = c("full", "fixed", "structural", "covariate", "random"),
  .show_desc = TRUE,
  .width = 1
)
}
\arguments{
\item{.df}{bootstrap parameter dataset and non-bootstrap parameter dataset combined.}

\item{.pmtype}{parameter table type. Options include:
\itemize{
\item \code{"full"} (all rows in \code{.df} retained in pmtable). This is the default.
\item \code{"fixed"} (all rows with type = "Struct" or "effect"),
\item \code{"structural"} (all rows with type = "Struct"),
\item \code{"covariate"} (all rows with type = "effect"),
\item \code{"random"} (all rows with greek = "Omega" or type = "Resid").
}}

\item{.show_desc}{logical (T/F). If \code{TRUE} (the default), include the description
column in the final table.}

\item{.width}{notes width. Defaults to 1.}
}
\description{
Generates specific parameter tables by filtering and using pmtables

This function:
\enumerate{
\item Filters to columns needed for specific parameter tables
\item Panels by "type"
\item Makes "abb", "greek", "desc" blank for "full", "fixed", "structural", "covariate" parameter tables.
Makes "abb" and "greek" blank for "random"
\item Attaches notes
\item Rename "value" to "Estimate" and "shrinkage" to "Shrinkage (\\\%)", if applicable
}

If these pmtable settings do not work for your parameter table, you can overwrite them afterwards using desired pmtables commands.
}
\examples{

\dontrun{

model_dir <- system.file("model/nonmem", package = "pmparams")
paramKey <-  file.path(model_dir, "pk-parameter-key-new.yaml")

# Parameter estimates
mod <- bbr::read_model(file.path(model_dir, "106"))
param_df <- define_param_table(
 .estimates = mod,
 .key = paramKey,
) \%>\% format_param_table()

# Bootstrap estimates
boot_run <- bbr::read_model(file.path(model_dir, "106-boot"))
boot_df <- define_boot_table(
 .boot_estimates = bbr::bootstrap_estimates(boot_run),
 .key = paramKey
) \%>\% format_boot_table()

# Combine parameter estimates with bootstrap estimates
combine_df <- left_join(param_df, boot_df)

# Fixed effects table
make_boot_pmtable(.df = combine_df, .pmtype = "fixed") \%>\%
 stable() \%>\%
 st_as_image(border = "0.8cm 0.7cm")


# Random effects table
make_boot_pmtable(.df = combine_df, .pmtype = "random") \%>\%
 stable() \%>\%
 st_as_image(border = "0.8cm 0.7cm")
}
}
