% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pmtable.R
\name{make_pmtable}
\alias{make_pmtable}
\title{Make pmtable parameter table}
\usage{
make_pmtable(
  .df,
  .pmtype = c("full", "fixed", "structural", "covariate", "random"),
  .width = 1
)
}
\arguments{
\item{.df}{parameter data set output from pmparams::format_param_table.}

\item{.pmtype}{parameter table type. Options include:
\itemize{
\item \code{"full"} (all rows in \code{.df} retained in pmtable). This is the default.
\item \code{"fixed"} (all rows with type = "Struct" or "effect"),
\item \code{"structural"} (all rows with type = "Struct"),
\item \code{"covariate"} (all rows with type = "effect"),
\item \code{"random"} (all rows with greek = "Omega" or type = "Resid").
}}

\item{.width}{notes width. Defaults to 1.}
}
\description{
Make pmtable parameter table
}
\details{
Generates specific parameter tables by filtering and using \code{pmtables}

This function:
\enumerate{
\item Filters to columns needed for specific parameter tables
\item Panels by "type"
\item Makes \code{"abb"}, \code{"greek"}, \code{"desc"} blank (no title)
\itemize{
\item Note that description is removed when \code{.pmtype = "random"}. See
\code{?pmtables::st_mutate()} if you want to add it back in.
}
\item Attaches notes
\item Rename "value" to "Estimate" and "shrinkage" to "Shrinkage (\%)", if
applicable
}

\strong{Note:}
If these \code{pmtables} settings do not work for your parameter table, you can
overwrite them afterwards using desired \code{pmtables} commands.
}
\examples{

model_dir <- system.file("model/nonmem", package = "pmparams")
paramKey <-  file.path(model_dir, "pk-parameter-key-new.yaml")

# Read in parameter estimates (or provide `bbr` model)
param_est <- utils::read.csv(file.path(model_dir, "param_est.csv"))

# Create formatted parameter table
param_df <- define_param_table(
 .estimates = param_est,
 .key = paramKey
) \%>\% format_param_table()

\dontrun{

# Create random effects table that can be rendered to latex
make_pmtable(param_df, .pmtype = "random") \%>\%
 pmtables::stable() \%>\%
 # preview in Rstudio viewer (requires `magick` and `pdftools`)
 pmtables::st_as_image()
}

}
\seealso{
\code{\link[=make_boot_pmtable]{make_boot_pmtable()}}
}
