% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_key.R
\name{param_key}
\alias{param_key}
\title{Parameter key requirements}
\description{
The parameter key file used in the mrgparamtab function must contain the following columns:

\strong{name} = THETA, OMEGA, SIGMA

\strong{abb} = abbreviation you want to appear in the parameter table (use latex coding)

\strong{desc} = description you want to appear in the parameter table

\strong{panel} = indicate which panel each parameter should appear under.

Current options include:
\itemize{
\item panel=="struct" ~ "Structural model parameters"
\item panel=="cov" ~ "Covariate effect parameters"
\item panel=="IIV" ~ "Interindividual covariance parameters" (off-diagonals << function takes care of this)
\item panel=="IIV" ~ "Interindividual variance parameters" (diagonals << function takes care of this)
\item panel=="IOV"  ~ "Interoccasion variance parameters"
\item panel=="RV" ~ "Residual variance"
}

Additional options can be added to \code{getPanelName} function in \code{functions-table.R}

\strong{trans} = define how the parameter will be transformed. Current options include:
\itemize{
\item "none" -   untransformed parameters, e.g. THETAs or off-diagonals
\item "logTrans" -  THETAs estimated in the log domain
\item "logitTrans" - THETAs estimated using a logit transform
\item "lognormalOm"- for log-normal omegas e.g. CL = THETA(1) * exp(ETA(1)) - returns est.+CV\%
\item "OmSD"       - for omegas where you want to return SD only - returns estimate & SD; use this for additive omegas e.g. CL = THETA(1) + ETA(1)
\item "logitOmSD"  - for omegas using logit transform - returns estimate & SD (calculated with logitnorm package); this option requires you provide the associated THETA separated with a "~"; e.g. "logitOmSD ~ THETA3"
\item "addErr"     - for additive error terms (coded using SIGMA in $ERROR) - returns est.+ SD
\item "propErr"    - for proportional error terms (coded using SIGMA in $ERROR) - returns est.+CV\%
}
}
