% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-notes.R
\name{param_notes}
\alias{param_notes}
\title{Generate all footnote equation options}
\usage{
param_notes(.ci = 95, .zscore = NULL)
}
\arguments{
\item{.ci}{specify 90 or 95 percent confidence interval (default 95\%)}

\item{.zscore}{z-score for the specified confidence interval. Only needed for
confidence intervals that are NOT 90 or 95 percent}
}
\description{
Generate data frame of generic footnote equations to append to parameter tables.
}
\examples{
library(dplyr)
model_dir <- system.file("model/nonmem", package = "pmparams")
paramKey <-  file.path(model_dir, "pk-parameter-key-new.yaml")

# Read in parameter estimates (or provide `bbr` model)
param_est <- utils::read.csv(file.path(model_dir, "param_est.csv"))

# Make and format parameter table
defineOut <- define_param_table(
 .estimates = param_est,
 .key = paramKey,
 .ci = 95
) \%>\% format_param_table()

# Make random effects table and add relevant footnotes:
footnotes <- param_notes()

table <- make_pmtable(.df = data, .pmtype = "random") \%>\%
  pmtables::st_notes(footnotes$ci, footnotes$cv) \%>\%
  pmtables::st_notes_str() \%>\%
  pmtables::st_notes(footnotes$cvOmegaEq, footnotes$cvSigmaEq)

}
