% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmparams.R
\name{pmparams}
\alias{pmparams}
\title{pmparams}
\description{
\code{pmparams} is a library written in R that generates clear,
well-formatted parameter tables to report NONMEM model results.
}
\details{
\code{pmparams} creates parameter tables by leveraging a few intuitive functions designed to:
\itemize{
\item Read in your model parameters from various NONMEM output files
\item Describe your model parameters in a YAML file
\item Back-transform any parameters estimated in other domains (e.g., log- or logit-transformed variables)
\item Calculate additional summary statistics (e.g., 95\% confidence intervals, coefficient of variation (CV), etc.)
}
}
\keyword{internal}
