% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removePunc.R
\name{removePunc}
\alias{removePunc}
\title{Remove punctuation from parameter names}
\usage{
removePunc(.df, .column)
}
\description{
The model output for parameter estimates include punctuation in the parameter
names, such as \code{OMEGA(1,1)}. Often it is desirable to have the name as
\code{OMEGA11} instead.
}
\keyword{internal}
